requirements.txt
flask==3.1.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.12
bcrypt==5.0.0
qrcode[pil]==8.2
pillow==12.2.0
gunicorn==25.3.0
python-dotenv==1.2.2
